LayoutTests/imported/w3c/web-platform-tests/custom-elements/parser/parser-sets-attributes-and-children-expected.txt


PASS HTML parser must set the attributes 
PASS HTML parser must append child nodes 
PASS HTML parser must set the attributes or append children before calling constructor 
FAIL HTML parser should call connectedCallback before appending child nodes. assert_equals: expected 0 but got 2
PASS HTML parser must enqueue attributeChanged reactions 
hello world
